feat: #1070 - refactored, in order to be closer to the previous version
Impacted files:
* `edit_ingredients_page.dart`
* `ocr_helper.dart`
